Fix broken logic in Consumer::ByTime->expire_date
The old logic was quite wrong and commingled charge frequencies and
cost-periods.  Since expiration will occur when the consumer attempts
to charge an underfunded bank, it must be calculated on the basis of
charge frequency.  The first step is to compute the cost per charge;
there is a method for this already.

Then dividing the banked amount by cost-per-charge gives the (possibly
fractional) number of charges that can be applied before the bank runs
out; the floor of this number is the number of full charges that can
be applied before the bank is underfunded for another charge.

If this number is zero, then the consumer will expire at the *next*
charge date; if it is N>0 then the consumer will expire at the next
charge date plus N.  (That is, there will be no expiration for N
charges, and the expiration will occur at the N+1th.)