CanTransfer role
If an object composes CanTransfer, it gains an immutable class
attribute, transferer_type, that identifies its type to the transfer
accountant.  For example, $bank->transferer_type is "bank".

Also, add a TransferCapable type to Moose::Types, which is the type of
all such identifiers; these are validated by
TransferUtil::is_transfer_capable, which accepts items listed in the
table in TransferUtil.