rephrase construct_own_replacement in terms of templates
In other words, consumers are responsible for knowing the *template* to use to
replace them, rather than for building the actual thing.  This means they'll
almost always be "pobox.mailstore" for example, but they can return a full
template, too.

Next up, make Moonpig::URIs that resolve to template names.