nearly a day's work on generalizing the router
...and it mostly sucks.

The router now crawls down routable objects, looking to find an invocable
object at the end -- which right now is only allowed to be a bogus
WrappedMethod object.

This needs to be put into a role so that we can end with plain ol' objects to
do things like GET an invoice or POST to a collection of consumers.

This whole hunk of work made me pretty miserable, and it should be reworked or
replaced with something more standard in the future, but I did this after
spending two days determining why all the stuff already out there was not going
to make us happy.