reverse order of ->become_active and ->terminate_service in Consumer->copy_to
When copying an active consumer from ledger A to B, we must terminate
the consumer in A before activating the copy in B, or else we fall
afoul of the constraint that says that only one ledger may have
an active consumer for a particular xid.

This depends on the previous commit, which makes Storage::Spike honor
the order in which changes are made to ledgers.

Also rename Consumer->move_bank_to to ->move_bank_to__, as it is an
internal-only method.