Banks are now created by Invoices, not by Charges
Formerly, banks were created by charges when they were paid.  But if a
consumer put multiple charges on an invoice, it was not clear which
one should create the bank, or for how much.  A consumer could have
only one bank, which was immutable, so the bank had to be created with
the correct total amount.

Reponsibility for creating the bank now rests with the invoice, which
processes its charges when it is paid and totals up the required
amount for each consumer.