Fix up move-consumer-across-ledgers code to work now that not all consumers have banks
the copy_to method now has a submethod, copy_subcomponents_to__, which
is responsible for copying any subsidiary objects of the main
consumer.  This includes the consumer's replacement, if it has one; in
consumers with banks, it includes the bank.  The ChargesBank role
decorates copy_subcomponents_to__ with code to copy over the bank.