New coupon type: bulk purchase discounts
Plus tests.  The interesting thing here is that although a regular
bulk discount is 10%, a nonprofit organization already gets an
across-the-board discount of 10%, and we don't want to apply the bulk
discount here to add up to a 19% total discount.  Rather,
bulk-purchase has a special case: when it applies to a nonprofit
account, it brings the total discount to 15%, not 19%.