test for coupon-ordering issue
Depending on how coupon application code works, it could be
significant what order the coupons are processed in.

For example, suppose there is a $100 charge and (A) a 10%-off and (B)
a $10-off coupon.  If A is applied and then B, the total discount will
be $20.  If B is applied and then A, the 10%-off might be calculated
on the basis of there being only $90 left to pay, and the total
discount would then be only $19.

At present, this does not occur: coupon values are always calculated
based on the charge amount, not on the amount left to pay.  This test
checks to make sure that is true.

Should it ever start to fail, some code to fix it is available at
mjd/coupons-ordering.