Allow some periodic-charge consumers to overdraw their banks
This is necesary to support certain special consumers, such as Bulk,
which represent large customers to whom we are willing to extend
credit.

The accountant's "create_transfer" method already has a flag
('skip_funds_check') you can supply that tells it to skip its usual
sufficient-funds check. Here we push that flag up though the ledger to
the journal's 'charge' method so that the consumer can set it when it
calls Journal->charge.

ChargesPeriodically also gets a new data member which says whether the
consumer is allowed to overdraw. Naturally, this defaults to false!