Storage->retrieve_ledger_for_guid returns unwritten modified ledger when possible
If you modify a ledger inside a transaction and then save it, the save
is deferred to the end of the transaction.  Calls to
storage->retrieve_ledger_for_guid would formerly return the unmodified
version from storage instead of the saved but unwritten modified
version.  This changes that behavior; the saved but unwritten object
is now returned directly from the save queue, if possible.