Rework build() returned hash as $ledger->get_component() method
Formerly testing was in terms of a build() function that would return
a 'stuff' hash populated with named ledger and consumer objects.  This
wasn't going to work if ledgers can be specified only via GUID,
because any consumer objects returned from the build call might not be
the objects that were actually inside the ledger inside a transaction.
(Say if the ledger was rebuilt from the pickled version.)

So the 'stuff' hash has been embedded inside the ledger now, and is
accessed via a $ledger->get-component method.

Also redid the docs to warn that build() may return invalid objects,
and to put do_with_ledger up front.