Eliminate EventHandler::Code
EventHandler::Code is problematic because it contains a
coderef. Storable can't save coderefs, so any ledger with such an
event handler cannot be saved. Formerly it was used only in testing,
in transient ledgers that were never saved.  But since the shift to
do-with-ledger, all ledgers are saved, or are subject to being saved,
so these tests failed, and EventHandler::Code seems to be no longer
practical.