bug fix in do_with_ledger_array
The old code did not attempt to pass the retrieved ledgers in the same
order that they were requested by the caller of do_with_ledger_array.
Instead the order was determined by hash key order on the internal
hash constructed for passing to do_with_ledgers.  This meant that when
N ledgers were passed, they would be passed in the wrong order
1-(1/N!) of the time.