Rework interface to do_with_ledger.
do_with_ledgers now takes an array, not a hash, and the called code
gets an arg list of ledgers, not a hash.

Also, moved all the $options hashes to the front of the arg list,
which necessitates checking to see if it was omitted.  For
do_with_fresh_ledger, also merged the options hash into the main build
argument.  It doesn't really belong there, but that argument it kind
of a potpourri anyway.

Also renamed do_with_test_ledger to do_with_fresh_ledger.