Split out invoice charge construction into a separate method
Formerly Consumer::InvoiceOnCreation would add charges to the invoice
itself, assuming that they should be of class InvoiceCharge::Bankable,
which might not be correct.

Now the charge arguments are passed to an intermediate required
method, ->build_charge, which is implemented elsewhere.  For the old
behavior, it is provided in Consumer::ChargesBank, which just uses
class InvoiceCharge::Bankable as before.  But for some consumers, it
might need to create a fancier charge, or to do something else also.