pogrom against bankers
Banks are gone.  Banks were originally fixed-value hunks of cash.  When your
bank was nearly exhausted, it was time to make a new consumer because *no
matter what* your bank was not going to grow any more money.

While working on implementing overdraft, we realized that it was going to be
very difficult -- or at least tedious and annoying -- to pay off the overdraft
against a bank on an expired consumer.  We also realized that we had a lack of
good paper trail between credits and banks, because this used to be the case:

1. credit is applied to invoice with transfers
2. bank is created without reference to the invoice or credit
3. invoice's funds-on-hands remain with it forever

We decided to eliminate the "amount" attribute, and instead create a new kind
of transfer, transfering funds from the invoice to the bank.  Now we had a
paper trail.  Further, we could transfer *more* funds to a bank later, from an
invoice that included an overdraft reconciliation charge.  This made testing
just a little more complicated, because we'd have to create transfers into the
bank to fund it, instead of creating it with funds ex nihilo.  So, we created a
new kind of transfer.

Now, banks were created with no funds.  They had no funds until something was
transferred to them.  Once they got funds, they still might get more.  What's
the point of a bank in that scenario?  It's just a thing to transfer money to
and from.  We could use the consumer for that, which not only has the benefit
of eliminating a whole class of objects (banks) but also of eliminating any
branching logic that had to look at `$consumer->has_bank`.