per-env share roots are now extra_share_roots
When Fauxbox::Moonpig::Env subclassed Env::Test, it provided share
roots, which it wanted to be "extra".  The problem was that the
advice that added the core roots had been composed into the core
method, and was replaced by the subclass method.  (Moose is not
CLOS: advice is lost by method overriding, it is not distinct from
methods.)

Now you can just replace extra_share_roots, if you want.