Events no longer suppressed for expired objects
This was a terrible idea. The semantics of expiration are not defined
for all possible objects, and even for consumers it is the wrong thing
to do. An expired consumer might receive further events that it needs
to handle, and should not ignore them; event processing should not be
dependent on the order of events in this way.

In fact, I think there is a bug in this version where an expired
consumer will be sent a "terminate" event, which it then erroneously
ignores.