Consumers and ledgers now ->estimate_cost_for_interval
And return the amount of money they expect to cost to run for a
specified anount of time.

This is needed so that if the customer gets an invoice for (say) a
year of service and wants to pay instead of two years of service, we
can get a reasaonable guess about what to charge them up front.

Each consumer will do this differently; this commit only commits code
for Consumer::ByTime.  Ledgers estimate costs by adding up the
estimates for all their active consumers.

Potential problems: Is a "year" equal to days(365) or days(365.25)?
Moonpig::Util::years says 365.25.  So if a consumer says its
cost_period is days(365), and then you ask for a cost estimate for
years(1), it will give you an amount that is 1/1461 higer than you
expected; if you were expecting $100, it will say $100.07.  Watch out.