add yearly cost estimate to packed version of ledger
WARNING: this implementation tells the ledger to ignore consumers that
don't implement their own cost estimates. The CORRECT implementation
is to throw an exception on such a condition.  But to make that work,
we would have to fix a bunch of tests that use raw class("Consumer")
objects, which cannot be made to do such estimates.