add the "pay_as_imported" arg to ledger creation
If passed and true, "pay_as_imported" will automatically create credits
for all the consumers that were created (and invoiced).  It uses
non-refundable credit for the currently-active consumers, and refundable
credit for the rest.  (It is not yet clear that this is sufficient to
get the right kind of refund behavior, but it seems at least part of it.)