cache or queue even if retrieved ledger was cached
Previously, we only queued a save if the ledger had to be retrieved
from the db.  We should queue a save no matter what.  On one hand,
we should never be able to start a new xact with a ledger already
cached but not queued (after fixing a bug a few commits ago, anyway)
but on the other hand, it would be better to re-queue unnecessarily
than to fail to save!