clear the cache after last transaction completes, even on exception
The cache must not remain populated just because the usual "clear the
cache" logic is bypassed by exceptional flow control.  The cache must
always be cleared when the topmost transaction is completed, even if
"completed" means "dies."

** PLEASE NOTE **

It is important that no transaction ever catch an exception thrown by
a nested transaction, then attempt to continue operating as normal.
The nested transaction may have left objects in an inconsistent state
that must not be written out to storage.

For example:

  do_rw(sub {
    try {
      do_rw(sub {
        $ledger->update_that_works;
        $ledger->update_that_dies;
      });
    } catch {
      warn "oh no! <<$_>>";
    };

    $ledger->some_other_stuff;
  });

Here, the outer transaction will still continue after the fatal error
thrown inside the nested transaction.  The changes made in
"update_that_works" will be written, even though they *must not be*
without the other related changes.

If a transaction dies, any objects handled by it must be thrown away
and not considered valid for inspecting, changing, or saving.