amend POST-created ledgers to get them cached
...otherwise this happens:

  my $ledger = class('Ledger')->new(...);
  $ledger->add_some_consumers_which_charge;
  $ledger->perform_dunning;

Dunning ends up looking for $invoice_charge->owner.  InvoiceCharge
objects don't have a ref to their owner or ledger.  They have guids.  So
to get the owner, they do:

  sub InvoiceCharge::ledger {
    Storage->retrieve_ledger_by_guid( $self->ledger_guid);
  }

  sub InvoiceCharge::owner {
    $self->ledger->consumer_by_guid( $self->owner_guid );
  }

During initial creation of ledgers via PostTarget, though, the ledger
wasn't in the cache, so the retrieve call looked in the db, didn't find
it, and chaos broke look: we tried to call a method on undef, because
the call to $self->ledger, above, returns undef.  (I'll add a sanity
check for that.)

So, anyway, the solution is to cache the ledger as soon as possible once
it is instantiated.