fix overzealous frozen clock restorer
Without this fix, if the clock is stopped before an in-process request
to the PSGI app, the request handler will start by restoring the time,
which resets to the "not" stored during db initialization.

I had worked around this by calling $storage->_store_time but did not
like using that private method to bypass the bug.  Instead, I use 0 as
the "never stored time" value so that we know whether to really restore
the stored time and offset.