move most of the charge execution logic to Invoice
This is something I knew would be needed as soon as the original
code was written.  Charges are not ledger components, and do not
have easy access to the ledger on which they appear, so it's not
trivial for them to get the ledger for things like creating
transfers.

Now, consumers find the invoices that have charges for them and
sends the invoice a "execute charges for $consumer" message.