update consumer's balance copying for jit-pay
We don't need to create an invoice, we can just do a direct
credit-to-consumer transfer.  This is a *much* better model for what
we're doing anyway.

The old invoice-creating code seemed likely to lead to weirdness anyway.
It created an invoice, but didn't close the existing invoice.  This
means that the last invoice in the invoice collection would no longer be
the current open invoice... except that this invoice didn't go in the
collection anyway.  It might have been garbage collected, but this would
leave dangling symbolic references in the accountant.

I stumbled over this for a while before realizing we could eliminate the
entire invoice-creation big here and do the transfer directly.