let ByTime consumers predict shortfalls
They figure out whether their expected duration is more than one charge
cycle less than their original proration_period.  If it is, they return
the duration by which it is less.

This is not implemented in terms of existing methods for various
reasons.  expiration_date is pessimistic about getting paid for
outstanding charges, but we want to be optimistic.  It also won't run on
inactive consumers.  estimated_lifetime assumes all funding will be
there, for inactive.  It goes on.

Probably the right thing is to decide whether we want the behavior now
in shortfall prediction to be ported back to the other prediction
methods, and then we can do that and reduce duplication all over.