Fix quotes for service extensions
Instead of a template consumer, these get a xid. The replacement chain
for the currently activated consumer is calculated but not attached,
and the chain is started from there.

When executed, the quote looks to see if service is currently active.
If it is, the quote's first consumer is installed as the replacement
for the currently-active consumer; if not, the quote's first consumer
is activated immediately.