allow quote_guid to be passed with credit addition
This is *CLEARLY* a bogus hack.  This is one of the places where we
clearly want the request multiplexer for Stick that we had talked about
in the past.  Still, we don't have it yet, and it's not going to get
designed and added and tested at this late date for deployment.  When we
add it later, this one stupid exception can be removed.