alternate fix for psynch invoice error
predicted_shortfall already counts chrges on closed but unpaid
invoices as funds that it expects to get when figuring how long it
will live.  This changes it to count charges on open invoices as well.

This is necessary because when importing a new account, the heartbeat
that will close the invoice is sent to the consumer first, and
triggers a premature psynch invoice.  An alternative fix would be to
have the ledger explicitly close the invoice, but I believe the
solution in this commit will be more robust. The same situation
(heartbeat triggers psynch invoice before it closes invoice) could
occur elsewhere, and the change here will defend against it.