ByTime consumers should make replace, at least if ever funded
Without this, a consumer that was funded, but not very much, it would
never be able to make a payment, meaning the "around" advice on
charge_one_day would return too early -- before the $orig was called,
which was the only place that maybe_make_replacement was called.

We added a maybe_make_replacement call to the advice, but it was calling
too aggressively -- replacements that were never paid and were destined
to be expired were making replacements ad infinitum!  We only want
something to make a replacement if it (a) got some money (b) kept some
of it.