allow a ByTime consumer to pack with zero cost
When a consumer has a $0 total charge, that is a problem and should prevent
many kinds of normal operation... but not packing for display!  If we can
not predict the lifetime because of a $0 cost, we should die, but if we
only hit that path during packing, we should just omit the data.