clear the event handler registry when saving
Wow!

We really shouldn't be saving the result of composing the implicit event
handlers.  In fact, the idea of implicit event handlers was specifically
so that we could fix them behind the scenes without screwing with the
objects.  The fact that we were saving them is a /huge/ stupid problem,
and was the reason that the recent s/charge/maybe_charge/ problem
happened.  We weren't charging maybe_charge, but charge, because the old
event handlers were being used.

This change is a hack:  we stop saving the registry.  That means we lose
any *explicit* event handlers, but that's okay, because right now we
only use them in testing.  In the near future, we will refactor,
storing explicit handlers on the objects themselves and recomposing them
lazily as needed.