make published "expiration_date" call a non-published method
This all needs to be reworked to put the publication in
PredictsExpiration, but for now this will work.  We want to be able to
override a NON-published method so we do not affect the published code.

We need to override this to allow a faked-up expiration date in some
scenarios for "zero cost" by time consumers.