it is okay to make a replacement if we never invoiced
In 7c7222b we started to make replacements as long as we'd ever been paid.
This is good, but sometimes we have a consumer whose initial invoice charges
total 0, meaning that it never would have issued an invoice or been funded.
These are used to track groups of complimentary accounts, in Pobox.  If one of
those accounts becomes non-comp, it needs to issue an invoice immediately,
which could happen normally if only it created a replacement -- so we need to
make an exception and allow a replacement to be made if we weren't funded
*because we never invoiced.*