move shortfall clipping into psyncer
_predicted_shortfall is a low-level routine, and shouldn't be applying
policy decisions like whether certain amounts of shortfall are
considered tolerable.  I turned it into a toutine that just returns
the amount of shortfall time, even if that amount is *negative*, which
indicates a surplus of funds.

Then _send_psync_quote is responsible for dealing with that
appropriately. If the shortfall is small, the psyncer ignores it.