Test fixes
Formerly the test changed the charge rate on the consumers from $1/day
to $2/day; I decided this was too likely to cause spurious test
successes because the shortfall amount matched the paid amount
exactly.  I changed it to $1.42857/day, which is unlikely to coincide
with anything else.

Fixed last test failure which was caused by a failure to throw away
the initial invoice.