Various fixes in psync tests
Set grace period small, since that was preventing charges  from
reducing the consumer's unapplied funds, and so was simplifying the
psync calculations inappropriately.

Advance the clock *after* posting a heartbeat in elapse(), so that the
first heartbeat is posted in the first day, instead of on the second.

Correct expected shortfalls and charge amounts.