correct the calculation of expected discount
We ended up off by one millicent, because of where integer truncation
was happening.  To match the coupon appplication, we truncate the
discount and subtract that from the charge, rather than truncating
((1 - discount%) * charge).  The error will only be a rounding error,
so just 1 millicent, but we can get it right, so let's.

We can always make this a "within 1 mc" in the future.