Insert expiration dates into psync notice
Dunner's send_psync_email method now expects to get an info hash that
contains the expiration dates, in addition to the raw quote; it is no
longer expected to infer the dates from the consumer and the quote
themselves.

The Consumer's maybe_send_psync_quote method now calculates the dates
and passes them to the Dunner.