cashout no more than the available-to-cashout amount
We can't assume that the whole applied amount of a credit is applied
to the consumer under consideration.  In fact, we already know how
much of it is -- it was part of the effective funding pairs that we got!

We should refund min( applied-to-this, amount-needed-to-refund ), which
is what this commit makes us do.