reset dbh per requeset
We have been seeing an occasional strange bug where Ledger A tries
to claim an xid from Ledger B.  When this happens, every ledger we
try to save after that will continue to due the same.  Ledger C, D,
E, F, etc., will all try to claim that xid.  Fortunately, there are
constraints on the database that prevent the update, but it means
that many transactions are lost.

On the theory that some strange DBI bug is to blame, we are making
a new connection per request.  We do not have evidence for there
being such a bug, but extensive investigation in our own code has
not uncovered anything.