Tinker with ByTime::_issue_psync_charges
It no longer issues a charge and then calls itself recursively on the
replacement chain of the current consumer.  Instead, the active
consumer is responsible for getting the chain and looping over the
descendants, asking each one to add its own charge.

This is so that B5G1 consumers can wrap the _issue_psync_charge method
to have a different behavior without requiring them to know how to
redispatch the call.