put the implicit handler definition AFTER with(Consumer)
Why?  Because it breaks stuff.  The event handlers are not registered.

Why?  I have no idea.  Neither does MJD.

This bug needs to be reduced to a simple test case and either fixed or
documented as correct-but-surprising behavior.  Furthermore, MJD and I
now suspect we can probably rip out the whole event handling system and
replace it with advice.  Maybe.

For now, the problem has been bandaged over.