Move charge type constraints out of type system
The type system was getting cumbersome because the implementation we
had required an Invoice to contain only InvoiceCharges, but now we
also want it to be able to contain LineItems.  All charge containers
now accept any kind of Charge, but also must provide an accepts_charge
method to validate charges before they are added.

Got rid of the now-useless InvoiceCharge and JournalCharge types.
(The types, not the roles.)