Fix $invoice->all_charges to omit line items
->all_items now gets all items include non-charges.
->has_charges and ->has_items similarly.

Charges now have an ->is_charge method that says whether they are
included in the result from ->all_charges.  Invoice and Journal
charges are, line items are not.

Also added a couple of greps to disregard abandoned charges in places
where I think they ought to have been but weren't.