Rename "Charge" roles to "ChargeLike"
Since they now apply to line items, which are not considered to be
Charges, although they are charge-like in that they can go into
invoices.

InvoiceCharge  and JournalCharge remain as they were.

->all_charges is now named correctly, since it was always all
charges and never all chargelikes.

HasCharges name does not change.

The "Charge" type in Moonpig::Types is now "ChargeLike".