HasCharges collect LineItems; charges are items
This is a big commit, but it seemed best to keep it all together,
rather than try too hard to break it into non-test-breaking pieces.

* HasCharges means the object has a set of LineItems
* charges are a kind of line item
* counts_toward_total is gone; is_abandoned is used directly

This clears up a lot of otherwise confusing code where a LineItem did
the Charge role, but $line_item->is_charge was false.  Now there is no
abstract Charge, only InvoiceCharge and JournalCharge, each of which is
a kind of line item.

All line items count toward the total, unless abandonable and abandoned.
Line items can also be zero-amount, or even negative, now, though, so we
can add notices and discount items to invoices.

The CouponCreator code has been removed.  It was also removed on another
branch, so I've removed it here rather than updating it to the new
LineItem code.