correct the way we prevent abandoning no-abandoned-charge invoices
We were ending up unable to abandon zero-charge invoices because they
lacked any non-abandoned charges.  Zero-charge invoices are (sadly)
unavoidable at the moment.  They occur, for example, when a zero-charge
invoice is the current invoice and a quote is begun.

This got tickled when we started abandoning invoices with no
unabandoned charges left on them.  Sometimes the charge count was zero,
and those should get abandoned.  Who wants to look at a zero-item
invoice?  Ideally, these will be garbage collected in the future.