only mark the current replacement superseded
Any previous replacement should've been marked superseded if it was
superseded.  Superseding anything that used to be a replacement can be
dangerous in certain (currently not really supported) cases.  For
example...

  HEAD -> REPLACEMENTS

  A -> B -> C -> D -> E -> F

We want to remove E from the chain, so we do something like:

  push @{ $E->_replacement_history }, undef;
  $D->replacement( $F );

When the second line replaces $E with $F as $D's replacement,
$E->mark_superseded is called, which (before this patch) calls
mark_superseded on every consumer that was ever $E's replacement,
including $F, meaning that the chain now effectively ends with $D.

Instead, mark_superseded should only affect the current replacement
chain.  This will allow the two lines above to be implemented as, say,
->splice_next_replacement