make FixedExpiration abstract, ::Required concrete
There is a *bizarre* composition bug in Moose, it seems, that allowed
FixedExpiration, which provides expiration_date, to compose with a
Pobox-specific role (Domain::ThumbTwiddler) which also provided
expiration_date, without a conflict.  This led to Bad Stuff.

I want to provide a version that has all the logic *except* for the
attribute for the fixed expiration date.