allow funded-or-unfunded fixed expiration expectations
This lets a FixedExpiration consumer provide a behavior to be expected
if it is paid or isn't paid.

More importantly, it lets the consumer return a date (as it usually
will) or a duration.  This is useful to say things like "+0 seconds if
unpaid" but even more useful for domain consumers which can estimate "+X
years if paid" on top of the current date.