allow SelfFunding consumers to "psync" charge non-quotes
Before this commit, psyncking would always create a new psync quote, so
it was safe to assert that if a SelfFunding (or other) consumer was
doing a psync charge, the current invoice better be a psync quote.

The intent, now, is going to be to for consumers to reinvoice if they
need to psync and haven't yet been paid.  That means that the b5g1
consumer might not find a psync quote onto which to put its adjustment
charge.  We want to allow this.