make the ChargeTodaysDate invoice some extra funds
Without this, it was invoicing $1 (for Jan 1) on creation, but never
paying, it because the "bank" option to the factory puts funds directly
into the consumer, rather than paying off any invoices.

As I was looking into code to reinvoice, rather than psync, when a
psync would affect a consumer with an unpaid invoice, this test started
to show up via diagnostics.  It wanted to reinvoice.

Why?  Because its invoice was not paid!

So, I took $1 from the bank and put it into a credit so that the invoice
for $1 and $499 more were sent to the consumer directly.  This seemed
like it was going to cause more trouble, and certainly wasn't as easy to
explain, so now the consumer just asks for an extra hundred bucks up
front.

(Finally, I had also considered having the consumer start on the 31st of
December so that the 1st, 2nd, 3rd, etc. of January would all fit within
the inital $31, but of course the first journal charge would be for $31,
meaning that we'd blow all the money at once.  By definition, this
consumer can't live more than a day without asking for extra up front.)

This test would've been easier to fix if we could use ByTime without
psync.  We'll still want that later for this kind of test, because the
shortfall really does change.