do not re-invoice if charges are paid
The existing test before reinvoicing was:

    unless ($self->was_ever_funded) {

That kept us from reinvoicing if the consumer had already been given its
funds by the ledger.  Specifically, if its charges were executed.
That's not good enough!  The charges can be paid and not yet executed,
and we don't want to reinvoice if the first one was paid.  We want to
say that the consumer must be waiting for money, so we say:

    if ($self->_has_unpaid_charges && ! $self->was_ever_funded) {

This isn't great, because what happens with a consumer that didn't issue
any charges?  It won't reinvoice because it doesn't have any unpaid
charges.

There are a few places where we've had to say this sort of "did I ever
get paid the right way" thing, and it's not great.