reinvoice instead of psyncing sometimes
When we want to psync, but we have open charges, reinvoice instead.  We
reinvoice the consumer, which will make a new invoice for anything where
it's useful, and then we do a psync charge, which should only cause
charges for things that couldn't reinvoice.

Because this means we can end up with psync charges on real invoice
rather than an easier-to-replace quote, we make sure to abandon any
unpaid moonpig.psync-tagged charges before starting the process.  That
way if they end up needing to be for less, they can be.

Note that we never refund money from a paid consumer because of this.
Also, we don't redistribute the paid consumers' lifetime over the unpaid
ones.

If we had two consumers like this:

  * HEAD: asked for $50, has $50
  * REPL: asked for $50, has $ 0

...then if they both start needing $100, we'll issue a new invoice
asking for $150:  $100 for REPL and $50 to make up the difference on
HEAD.

...but if they both start needing $25, we'll issue a new invoice asking
for $25 to pay for REPL.  We won't transfer the overage from HEAD to
REPL.