update mail expectations in t/consumer/bytime/replacement.t
This had used the "bank more money than needed, but do not pay invoice"
strategy, which meant that we kept sending invoices for the "unpaid"
consumer even though it was funded.  I've rewritten these tests to
invoice for the amount of time/money desired, and then used
pay_payable_invoices to pay them.