when canceling (and thus terminating) abandon charges
Well, abandon *funding* charges!

The issue is this:

1. customer signs up for service
2. a year later, the customer gets a bill for new service and says,
"actually, I don't want to re-up" and instead of just waiting for
expiry, goes to terminate service
3. termination triggers a terminate event, which triggers a cancel
event and then sets the consumer inactive
4. ...but the bill is still there, and keep dunning "forever"
5. customer gets angry and will never recommend us to anyone again

A consumer can die in one of three ways:  expiration, cancelation, or
termination.  Expiration means "ran out of money with no replacement."
Cancelation means "set replacement to /nothing, which guarantees
expiration rather than replacement activation."  Termination means
"cancel and then become inactive (but do not expire)."

When we expire, right now, we abandon funding charges.  (That is, all
the charges that things filed that aren't tagged as "keep even after
expiring," which are useful for overage charges.)  We should do this for
cancelation, too.  If someone gets a bill and cancels rather than pay
it, we should stop asking them to pay it.  Bills are paid *before*
service starts, so this should be right.