when re-invoicing on rate change, do not psync
When there's a shortfall and the tail consumers on a chain are
entirely unpaid, we re-invoice so they can replace the old charges
(at the old rate) with new charges (at the new rate).  This avoids
the confusing situation of a user ending up with an invoice (at the
old rate) plus a psync quote (to keep the original intended
expiration date).

Unfortunately, this can happen:

  * consumer C, head of chain, experiencing shortfall of 90%
  * consumer D, C's replacement, never paid, with shortfall

D re-invoices, and then C issues its shortfall charges *onto the
invoice*, meaning that the charge must be paid to pay off D.  Nobody
wants this.

We could re-invoice D and then have C charge onto a quote, but the
user would then receive a quote and invoice, which is the initial
confusing situation we wanted to avoid.  It would be possible to
detect this situation when dunning and present a better-crafted
email, but that's More Work.

So, for now, when D re-invoices, C will not issue psync charges.