don't mark invoice closed if we can't cover their amount
There was a terrible bug here, effectively a caching bug.  We
computed the amount available, but then didn't decrease it as we
paid off invoices.  That means that all invoices at or below the
initially available amount would be marked paid.

Although this was very bad, it was not a complete nightmare.  This
would not lead to free service.  The invoices would be unable to
execute their charges properly, and the ledger would end up in an
over-earmarked state.  The amount needed would still show up as
"due," and the consumers affected would expire when unfunded.

Still, a nasty bug and one fixed by this commit.