one-off consumers!
These consumers issue one charge.  When it is paid, they die.  Further,
the consumers don't get or consume the funds paid to their charge.
Instead, the funds are effectively unconsumed, and get used by and other
service that needs them, later.

In other words, a OneOff consumer lets you ask for money to be sent in, later
to be used as "general funds."