Add same_day_next_month() for consistent monthly billing
This will get us to the same day the next month from now, rather
than just 30 days in the future.

If the current date is the 29th, 30th, or 31st we will instead go
to the first of the month *after* next month - since these days
don't exist on all months.