If we crash saving a ledger, clear our ledger queue
That way, we don't try to save it again the next time we work on
*any random ledger*.

Otherwise, what might happen is another process may save the ledger
we just tried to save. Then, when we pick up a new job to work on,
we have the out-dated version of that ledger in our _ledger_queue,
which will be saved at the end of our transaction along with
whatever ledger we're actually trying to work on.